add trading record model and fix id to uint64

+++ models/account.go
@@ -3,9 +3,9 @@ package models
 // Account ...
 type Account struct {
 	BaseModel
-	UserID     uint
+	UserID     uint64
 	User       User
-	CurrencyID uint
+	CurrencyID uint64
 	Currency   Currency
 	Balance    float64 `json:"balance" sql:"DECIMAL(32,16)"`
 	Locked     float64 `json:"locked" sql:"DECIMAL(32,16)"`
